Moved Profile structure to structures/profile.py
